Added modal to save model results
