+++ src/components/Modals/SaveModelResultsModal/SaveModelResultsModal.jsx
@@ -1,0 +1,16 @@
+import {Modal} from "../Modal";
+import {Button} from "../../Button/Button";
+import React from "react";
+
+export const SaveModelResultsModal = ({closeModal, open}) => {
+  return (
+    <Modal open={open} onClose={closeModal}>
+      <Button
+        text="Aceptar"
+        onClick={() => {
+          closeModal();
+        }}
+      />
+    </Modal>
+  );
+};
